Reduced apk size Theming name changed Cleanup Navigation improvements other

+++ terminal-emulator/src/main/java/com/termux/terminal/KeyHandler.java
@@ -20,77 +20,77 @@ public final class KeyHandler {
     static {
         // terminfo: http://pubs.opengroup.org/onlinepubs/7990989799/xcurses/terminfo.html
         // termcap: http://man7.org/linux/man-pages/man5/termcap.5.html
-        TERMCAP_TO_KEYCODE.put("%i", KEYMOD_SHIFT | KeyEvent.KEYCODE_DPAD_RIGHT);
+        TERMCAP_TO_KEYCODE.put("%i", Integer.valueOf(KEYMOD_SHIFT | KeyEvent.KEYCODE_DPAD_RIGHT));
         // Shifted home
-        TERMCAP_TO_KEYCODE.put("#2", KEYMOD_SHIFT | KeyEvent.KEYCODE_MOVE_HOME);
-        TERMCAP_TO_KEYCODE.put("#4", KEYMOD_SHIFT | KeyEvent.KEYCODE_DPAD_LEFT);
+        TERMCAP_TO_KEYCODE.put("#2", Integer.valueOf(KEYMOD_SHIFT | KeyEvent.KEYCODE_MOVE_HOME));
+        TERMCAP_TO_KEYCODE.put("#4", Integer.valueOf(KEYMOD_SHIFT | KeyEvent.KEYCODE_DPAD_LEFT));
         // Shifted end key
-        TERMCAP_TO_KEYCODE.put("*7", KEYMOD_SHIFT | KeyEvent.KEYCODE_MOVE_END);
-        TERMCAP_TO_KEYCODE.put("k1", KeyEvent.KEYCODE_F1);
-        TERMCAP_TO_KEYCODE.put("k2", KeyEvent.KEYCODE_F2);
-        TERMCAP_TO_KEYCODE.put("k3", KeyEvent.KEYCODE_F3);
-        TERMCAP_TO_KEYCODE.put("k4", KeyEvent.KEYCODE_F4);
-        TERMCAP_TO_KEYCODE.put("k5", KeyEvent.KEYCODE_F5);
-        TERMCAP_TO_KEYCODE.put("k6", KeyEvent.KEYCODE_F6);
-        TERMCAP_TO_KEYCODE.put("k7", KeyEvent.KEYCODE_F7);
-        TERMCAP_TO_KEYCODE.put("k8", KeyEvent.KEYCODE_F8);
-        TERMCAP_TO_KEYCODE.put("k9", KeyEvent.KEYCODE_F9);
-        TERMCAP_TO_KEYCODE.put("k;", KeyEvent.KEYCODE_F10);
-        TERMCAP_TO_KEYCODE.put("F1", KeyEvent.KEYCODE_F11);
-        TERMCAP_TO_KEYCODE.put("F2", KeyEvent.KEYCODE_F12);
-        TERMCAP_TO_KEYCODE.put("F3", KEYMOD_SHIFT | KeyEvent.KEYCODE_F1);
-        TERMCAP_TO_KEYCODE.put("F4", KEYMOD_SHIFT | KeyEvent.KEYCODE_F2);
-        TERMCAP_TO_KEYCODE.put("F5", KEYMOD_SHIFT | KeyEvent.KEYCODE_F3);
-        TERMCAP_TO_KEYCODE.put("F6", KEYMOD_SHIFT | KeyEvent.KEYCODE_F4);
-        TERMCAP_TO_KEYCODE.put("F7", KEYMOD_SHIFT | KeyEvent.KEYCODE_F5);
-        TERMCAP_TO_KEYCODE.put("F8", KEYMOD_SHIFT | KeyEvent.KEYCODE_F6);
-        TERMCAP_TO_KEYCODE.put("F9", KEYMOD_SHIFT | KeyEvent.KEYCODE_F7);
-        TERMCAP_TO_KEYCODE.put("FA", KEYMOD_SHIFT | KeyEvent.KEYCODE_F8);
-        TERMCAP_TO_KEYCODE.put("FB", KEYMOD_SHIFT | KeyEvent.KEYCODE_F9);
-        TERMCAP_TO_KEYCODE.put("FC", KEYMOD_SHIFT | KeyEvent.KEYCODE_F10);
-        TERMCAP_TO_KEYCODE.put("FD", KEYMOD_SHIFT | KeyEvent.KEYCODE_F11);
-        TERMCAP_TO_KEYCODE.put("FE", KEYMOD_SHIFT | KeyEvent.KEYCODE_F12);
+        TERMCAP_TO_KEYCODE.put("*7", Integer.valueOf(KEYMOD_SHIFT | KeyEvent.KEYCODE_MOVE_END));
+        TERMCAP_TO_KEYCODE.put("k1", Integer.valueOf(KeyEvent.KEYCODE_F1));
+        TERMCAP_TO_KEYCODE.put("k2", Integer.valueOf(KeyEvent.KEYCODE_F2));
+        TERMCAP_TO_KEYCODE.put("k3", Integer.valueOf(KeyEvent.KEYCODE_F3));
+        TERMCAP_TO_KEYCODE.put("k4", Integer.valueOf(KeyEvent.KEYCODE_F4));
+        TERMCAP_TO_KEYCODE.put("k5", Integer.valueOf(KeyEvent.KEYCODE_F5));
+        TERMCAP_TO_KEYCODE.put("k6", Integer.valueOf(KeyEvent.KEYCODE_F6));
+        TERMCAP_TO_KEYCODE.put("k7", Integer.valueOf(KeyEvent.KEYCODE_F7));
+        TERMCAP_TO_KEYCODE.put("k8", Integer.valueOf(KeyEvent.KEYCODE_F8));
+        TERMCAP_TO_KEYCODE.put("k9", Integer.valueOf(KeyEvent.KEYCODE_F9));
+        TERMCAP_TO_KEYCODE.put("k;", Integer.valueOf(KeyEvent.KEYCODE_F10));
+        TERMCAP_TO_KEYCODE.put("F1", Integer.valueOf(KeyEvent.KEYCODE_F11));
+        TERMCAP_TO_KEYCODE.put("F2", Integer.valueOf(KeyEvent.KEYCODE_F12));
+        TERMCAP_TO_KEYCODE.put("F3", Integer.valueOf(KEYMOD_SHIFT | KeyEvent.KEYCODE_F1));
+        TERMCAP_TO_KEYCODE.put("F4", Integer.valueOf(KEYMOD_SHIFT | KeyEvent.KEYCODE_F2));
+        TERMCAP_TO_KEYCODE.put("F5", Integer.valueOf(KEYMOD_SHIFT | KeyEvent.KEYCODE_F3));
+        TERMCAP_TO_KEYCODE.put("F6", Integer.valueOf(KEYMOD_SHIFT | KeyEvent.KEYCODE_F4));
+        TERMCAP_TO_KEYCODE.put("F7", Integer.valueOf(KEYMOD_SHIFT | KeyEvent.KEYCODE_F5));
+        TERMCAP_TO_KEYCODE.put("F8", Integer.valueOf(KEYMOD_SHIFT | KeyEvent.KEYCODE_F6));
+        TERMCAP_TO_KEYCODE.put("F9", Integer.valueOf(KEYMOD_SHIFT | KeyEvent.KEYCODE_F7));
+        TERMCAP_TO_KEYCODE.put("FA", Integer.valueOf(KEYMOD_SHIFT | KeyEvent.KEYCODE_F8));
+        TERMCAP_TO_KEYCODE.put("FB", Integer.valueOf(KEYMOD_SHIFT | KeyEvent.KEYCODE_F9));
+        TERMCAP_TO_KEYCODE.put("FC", Integer.valueOf(KEYMOD_SHIFT | KeyEvent.KEYCODE_F10));
+        TERMCAP_TO_KEYCODE.put("FD", Integer.valueOf(KEYMOD_SHIFT | KeyEvent.KEYCODE_F11));
+        TERMCAP_TO_KEYCODE.put("FE", Integer.valueOf(KEYMOD_SHIFT | KeyEvent.KEYCODE_F12));
         // backspace key
-        TERMCAP_TO_KEYCODE.put("kb", KeyEvent.KEYCODE_DEL);
+        TERMCAP_TO_KEYCODE.put("kb", Integer.valueOf(KeyEvent.KEYCODE_DEL));
         // terminfo=kcud1, down-arrow key
-        TERMCAP_TO_KEYCODE.put("kd", KeyEvent.KEYCODE_DPAD_DOWN);
-        TERMCAP_TO_KEYCODE.put("kh", KeyEvent.KEYCODE_MOVE_HOME);
-        TERMCAP_TO_KEYCODE.put("kl", KeyEvent.KEYCODE_DPAD_LEFT);
-        TERMCAP_TO_KEYCODE.put("kr", KeyEvent.KEYCODE_DPAD_RIGHT);
+        TERMCAP_TO_KEYCODE.put("kd", Integer.valueOf(KeyEvent.KEYCODE_DPAD_DOWN));
+        TERMCAP_TO_KEYCODE.put("kh", Integer.valueOf(KeyEvent.KEYCODE_MOVE_HOME));
+        TERMCAP_TO_KEYCODE.put("kl", Integer.valueOf(KeyEvent.KEYCODE_DPAD_LEFT));
+        TERMCAP_TO_KEYCODE.put("kr", Integer.valueOf(KeyEvent.KEYCODE_DPAD_RIGHT));
         // K1=Upper left of keypad:
         // t_K1 <kHome> keypad home key
         // t_K3 <kPageUp> keypad page-up key
         // t_K4 <kEnd> keypad end key
         // t_K5 <kPageDown> keypad page-down key
-        TERMCAP_TO_KEYCODE.put("K1", KeyEvent.KEYCODE_MOVE_HOME);
-        TERMCAP_TO_KEYCODE.put("K3", KeyEvent.KEYCODE_PAGE_UP);
-        TERMCAP_TO_KEYCODE.put("K4", KeyEvent.KEYCODE_MOVE_END);
-        TERMCAP_TO_KEYCODE.put("K5", KeyEvent.KEYCODE_PAGE_DOWN);
-        TERMCAP_TO_KEYCODE.put("ku", KeyEvent.KEYCODE_DPAD_UP);
+        TERMCAP_TO_KEYCODE.put("K1", Integer.valueOf(KeyEvent.KEYCODE_MOVE_HOME));
+        TERMCAP_TO_KEYCODE.put("K3", Integer.valueOf(KeyEvent.KEYCODE_PAGE_UP));
+        TERMCAP_TO_KEYCODE.put("K4", Integer.valueOf(KeyEvent.KEYCODE_MOVE_END));
+        TERMCAP_TO_KEYCODE.put("K5", Integer.valueOf(KeyEvent.KEYCODE_PAGE_DOWN));
+        TERMCAP_TO_KEYCODE.put("ku", Integer.valueOf(KeyEvent.KEYCODE_DPAD_UP));
         // termcap=kB, terminfo=kcbt: Back-tab
-        TERMCAP_TO_KEYCODE.put("kB", KEYMOD_SHIFT | KeyEvent.KEYCODE_TAB);
+        TERMCAP_TO_KEYCODE.put("kB", Integer.valueOf(KEYMOD_SHIFT | KeyEvent.KEYCODE_TAB));
         // terminfo=kdch1, delete-character key
-        TERMCAP_TO_KEYCODE.put("kD", KeyEvent.KEYCODE_FORWARD_DEL);
+        TERMCAP_TO_KEYCODE.put("kD", Integer.valueOf(KeyEvent.KEYCODE_FORWARD_DEL));
         // non-standard shifted arrow down
-        TERMCAP_TO_KEYCODE.put("kDN", KEYMOD_SHIFT | KeyEvent.KEYCODE_DPAD_DOWN);
+        TERMCAP_TO_KEYCODE.put("kDN", Integer.valueOf(KEYMOD_SHIFT | KeyEvent.KEYCODE_DPAD_DOWN));
         // terminfo=kind, scroll-forward key
-        TERMCAP_TO_KEYCODE.put("kF", KEYMOD_SHIFT | KeyEvent.KEYCODE_DPAD_DOWN);
-        TERMCAP_TO_KEYCODE.put("kI", KeyEvent.KEYCODE_INSERT);
-        TERMCAP_TO_KEYCODE.put("kN", KeyEvent.KEYCODE_PAGE_UP);
-        TERMCAP_TO_KEYCODE.put("kP", KeyEvent.KEYCODE_PAGE_DOWN);
+        TERMCAP_TO_KEYCODE.put("kF", Integer.valueOf(KEYMOD_SHIFT | KeyEvent.KEYCODE_DPAD_DOWN));
+        TERMCAP_TO_KEYCODE.put("kI", Integer.valueOf(KeyEvent.KEYCODE_INSERT));
+        TERMCAP_TO_KEYCODE.put("kN", Integer.valueOf(KeyEvent.KEYCODE_PAGE_UP));
+        TERMCAP_TO_KEYCODE.put("kP", Integer.valueOf(KeyEvent.KEYCODE_PAGE_DOWN));
         // terminfo=kri, scroll-backward key
-        TERMCAP_TO_KEYCODE.put("kR", KEYMOD_SHIFT | KeyEvent.KEYCODE_DPAD_UP);
+        TERMCAP_TO_KEYCODE.put("kR", Integer.valueOf(KEYMOD_SHIFT | KeyEvent.KEYCODE_DPAD_UP));
         // non-standard shifted up
-        TERMCAP_TO_KEYCODE.put("kUP", KEYMOD_SHIFT | KeyEvent.KEYCODE_DPAD_UP);
-        TERMCAP_TO_KEYCODE.put("@7", KeyEvent.KEYCODE_MOVE_END);
-        TERMCAP_TO_KEYCODE.put("@8", KeyEvent.KEYCODE_NUMPAD_ENTER);
+        TERMCAP_TO_KEYCODE.put("kUP", Integer.valueOf(KEYMOD_SHIFT | KeyEvent.KEYCODE_DPAD_UP));
+        TERMCAP_TO_KEYCODE.put("@7", Integer.valueOf(KeyEvent.KEYCODE_MOVE_END));
+        TERMCAP_TO_KEYCODE.put("@8", Integer.valueOf(KeyEvent.KEYCODE_NUMPAD_ENTER));
     }
 
     static String getCodeFromTermcap(String termcap, boolean cursorKeysApplication, boolean keypadApplication) {
         Integer keyCodeAndMod = TERMCAP_TO_KEYCODE.get(termcap);
         if (keyCodeAndMod == null)
             return null;
-        int keyCode = keyCodeAndMod;
+        int keyCode = keyCodeAndMod.intValue();
         int keyMod = 0;
         if ((keyCode & KEYMOD_SHIFT) != 0) {
             keyMod |= KEYMOD_SHIFT;
